backend/requirements.txt
Pillow==10.0.0
fastapi==0.100.0
uvicorn==0.23.1
python-multipart==0.0.6
faiss-cpu==1.7.4
numpy==1.24.3
pytest==7.4.0
python-dotenv==1.0.0
kaggle==1.5.16
tqdm==4.66.1
kagglehub==0.1.4
